+++ src/main/java/net/commoble/exmachina/api/content/DefaultSignalComponent.java
@@ -5,7 +5,6 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import com.mojang.serialization.MapCodec;